fix: circular import
- move function to other modulo to prenvent circular import
- fix por value on redis
